+++ ElectronicVotingSystem/src/main/java/com/cg/votingapp/dao/UpdateCandidateDao.java
@@ -7,6 +7,6 @@ public interface UpdateCandidateDao {
 	public CandidateEntity updateCandidateName(int candidate_id, String candidate_name) throws NullValueFoundException ;
 	public CandidateEntity updateAddress(int candidate_id, String address) throws NullValueFoundException ;
 	public CandidateEntity updateAge(int candidate_id, int age) throws NullValueFoundException ;
-	public CandidateEntity updateContactDetails(int candidate_id, int contact_details) ;
+	public CandidateEntity updateContactDetails(int candidate_id, long contact_details) ;
 
 }
